fix: Solucionar respuesta taskListId null para taskEntity

+++ src/main/java/com/lucas/tasktracker/dtos/responses/SimpleResponseTaskListDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class SimpleResponseTaskListDTO {
-    private int taskListId;
+    private Long taskListId;
     private String name;
 }
